// generated/src/app/api-models/ndc-code-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class NdcCodeMaster {
  ndcCode: string;
  description: string;
  effectiveDate: Date;
  terminateDate: Date;
  avgWholesalePrice: number;
  drugClass: string;
  codeQualifier: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}